requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1